MCH.ML/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv==0.19.0
psycopg2
requests
numpy
transformers
scikit-learn
torch
torchvision
img2vec_pytorch
tqdm
pillow
pandas
hnswlib
pickle